+++ src/ImageCache.hpp
@@ -4,6 +4,7 @@
 #define __IMAGECACHE_H
 
 #include <Geode/Geode.hpp>
+#include "utils.hpp"
 
 using namespace geode::prelude;
 
@@ -11,13 +12,15 @@ class ImageCache {
 
 protected:
     static ImageCache* instance;
+private:
+    std::string getKey(std::string key, std::string url);
 public:
 
     Ref<CCDictionary> m_imageDict;
 
     ImageCache();
-    void addImage(CCImage* image, std::string key);
-    CCImage* getImage(std::string key);
+    void addImage(CCImage* image, std::string key, std::string url);
+    CCImage* getImage(std::string key, std::string url);
 
     static ImageCache* get(){